auth-service/tests/requirements.txt
aiohttp==3.8.6
pydantic==2.7.1
pydantic_settings==2.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.3